third_party/WebKit/LayoutTests/fast/viewport/viewport-69-expected.txt

CONSOLE ERROR: line 3: The key "xinitial-scale" is not recognized and ignored.
ALERT: viewport size 100x110 scale 3.2 with limits [3.2, 5] and userScalable true
